+++ Goose/Commands/AetherCommand.cs
@@ -3,7 +3,7 @@ namespace Goose.Commands
     [Command("/aether ", Section = "General", Help = "Set the aether threshold.")]
     public sealed class AetherCommand : BaseCommand
     {
-        public void Execute(CommandContext ctx, decimal thres)
+        public void Execute(CommandContext ctx, double thres)
         {
             ctx.Player.AetherThreshold = thres;
         }
